fix handling notFound

+++ src/controllers/banner.controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interfaces";
 import BannerModel, { bannerDAO, TypeBanner } from "../models/banner.model";
 import response from "../utils/response";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 export default {
     async create(req: IReqUser, res: Response) {
@@ -56,6 +56,10 @@ export default {
         try {
             const { id } = req.params;
 
+            if(!isValidObjectId(id)) {
+                return response.notFound(res, 'failed find one a banner');
+            }
+
             const result = await BannerModel.findById(id);
 
             if(!result) {
@@ -71,6 +75,10 @@ export default {
         try {
             const { id } = req.params;
 
+            if(!isValidObjectId(id)) {
+                return response.notFound(res, 'failed update a banner');
+            }
+
             const result = await BannerModel.findByIdAndUpdate(id, req.body, {
                 new: true
             });
@@ -83,6 +91,10 @@ export default {
     async remove(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if(!isValidObjectId(id)) {
+                return response.notFound(res, 'failed remove a banner');
+            }
 
             const result = await BannerModel.findByIdAndDelete(id, {
                 new: true
