+++ src/controllers/category.controller.ts
@@ -2,6 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interfaces";
 import CategoryModel, { categoryDAO } from "../models/category.model";
 import response from "../utils/response";
+import { isValidObjectId } from "mongoose";
 
 
 export default {
@@ -60,6 +61,10 @@ export default {
         try {
             const { id } = req.params;
 
+            if(!isValidObjectId(id)) {
+                return response.notFound(res, 'failed find one a category');
+            }
+
             const result = await CategoryModel.findById(id);
 
              if(!result) {
@@ -75,6 +80,10 @@ export default {
         try {
             const { id } = req.params;
 
+            if(!isValidObjectId(id)) {
+                return response.notFound(res, 'failed update a category');
+            }
+
             const result = await CategoryModel.findByIdAndUpdate(id, req.body, {
                 new: true
             });
@@ -87,6 +96,10 @@ export default {
     async remove(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if(!isValidObjectId(id)) {
+                return response.notFound(res, 'failed remove a category');
+            }
 
             const result = await CategoryModel.findByIdAndDelete(id, {
                 new: true
