+++ src/controllers/ticket.controller.ts
@@ -58,6 +58,10 @@ export default {
         try {
             const { id } = req.params;
 
+            if(!isValidObjectId(id)) {
+                return response.notFound(res, 'failed find one a ticket');
+            }
+
             const result = await TicketModel.findById(id);
 
              if(!result) {
@@ -73,6 +77,10 @@ export default {
         try {
             const { id } = req.params;
 
+            if(!isValidObjectId(id)) {
+                return response.notFound(res, 'failed update a ticket');
+            }
+
             const result = await TicketModel.findByIdAndUpdate(id, req.body, {
                 new: true
             });
@@ -85,6 +93,10 @@ export default {
     async remove(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if(!isValidObjectId(id)) {
+                return response.notFound(res, 'failed remove a ticket');
+            }
 
             const result = await TicketModel.findByIdAndDelete(id, {
                 new: true
